add if around backup script output

+++ .zshrc
@@ -56,5 +56,8 @@ photoup() {
 }
 
 export PATH="$PATH:/home/mthorning/.cargo/bin"
-cat ~/.backup-status
-rm ~/.backup-status
+
+if [ -f "$FILE" ]; then
+    cat ~/.backup-status
+    rm ~/.backup-status
+fi
